+++ presto-main/src/main/java/com/facebook/presto/operator/AriaHashBuilderOperator.java
@@ -1,0 +1,293 @@
+/*
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.facebook.presto.operator;
+
+import com.facebook.presto.memory.context.LocalMemoryContext;
+import com.facebook.presto.spi.Page;
+import com.facebook.presto.sql.gen.JoinFilterFunctionCompiler.JoinFilterFunctionFactory;
+import com.google.common.annotations.VisibleForTesting;
+import com.google.common.io.Closer;
+import com.google.common.util.concurrent.ListenableFuture;
+
+import javax.annotation.concurrent.ThreadSafe;
+
+import java.io.IOException;
+import java.util.List;
+import java.util.Optional;
+import java.util.OptionalInt;
+
+import static com.google.common.base.Preconditions.checkState;
+import static com.google.common.base.Verify.verify;
+import static java.util.Objects.requireNonNull;
+
+@ThreadSafe
+public class AriaHashBuilderOperator
+        implements Operator
+{
+    @VisibleForTesting
+    public enum State
+    {
+        /**
+         * Operator accepts input
+         */
+        CONSUMING_INPUT,
+
+        /**
+         * LookupSource has been built and passed on
+         */
+        LOOKUP_SOURCE_BUILT,
+
+        /**
+         * No longer needed
+         */
+        CLOSED
+    }
+
+    private final OperatorContext operatorContext;
+    private final LocalMemoryContext localUserMemoryContext;
+    private final LocalMemoryContext localRevocableMemoryContext;
+    private final PartitionedLookupSourceFactory lookupSourceFactory;
+    private final ListenableFuture<?> lookupSourceFactoryDestroyed;
+    private final int partitionIndex;
+
+    private final List<Integer> outputChannels;
+    private final List<Integer> hashChannels;
+    private final OptionalInt preComputedHashChannel;
+    private final Optional<JoinFilterFunctionFactory> filterFunctionFactory;
+    private final Optional<Integer> sortChannel;
+    private final List<JoinFilterFunctionFactory> searchFunctionFactories;
+
+    private final PagesIndex index;
+
+    private State state = State.CONSUMING_INPUT;
+    private Optional<ListenableFuture<?>> lookupSourceNotNeeded = Optional.empty();
+
+    private Optional<Runnable> finishMemoryRevoke = Optional.empty();
+
+    private final LayoutSpecificAriaHash layoutSpecificAriaHash;
+
+    public AriaHashBuilderOperator(
+            OperatorContext operatorContext,
+            PartitionedLookupSourceFactory lookupSourceFactory,
+            int partitionIndex,
+            List<Integer> outputChannels,
+            List<Integer> hashChannels,
+            OptionalInt preComputedHashChannel,
+            Optional<JoinFilterFunctionFactory> filterFunctionFactory,
+            Optional<Integer> sortChannel,
+            List<JoinFilterFunctionFactory> searchFunctionFactories,
+            int expectedPositions,
+            PagesIndex.Factory pagesIndexFactory)
+    {
+        requireNonNull(pagesIndexFactory, "pagesIndexFactory is null");
+
+        this.operatorContext = operatorContext;
+        this.partitionIndex = partitionIndex;
+        this.filterFunctionFactory = filterFunctionFactory;
+        this.sortChannel = sortChannel;
+        this.searchFunctionFactories = searchFunctionFactories;
+        this.localUserMemoryContext = operatorContext.localUserMemoryContext();
+        this.localRevocableMemoryContext = operatorContext.localRevocableMemoryContext();
+
+        this.index = pagesIndexFactory.newPagesIndex(lookupSourceFactory.getTypes(), expectedPositions);
+        this.lookupSourceFactory = lookupSourceFactory;
+        lookupSourceFactoryDestroyed = lookupSourceFactory.isDestroyed();
+
+        this.outputChannels = outputChannels;
+        this.hashChannels = hashChannels;
+        this.preComputedHashChannel = preComputedHashChannel;
+
+        HashCollisionsCounter hashCollisionsCounter = new HashCollisionsCounter(operatorContext);
+        operatorContext.setInfoSupplier(hashCollisionsCounter);
+
+        layoutSpecificAriaHash = new LayoutSpecificAriaHash(hashChannels, outputChannels);
+    }
+
+    @Override
+    public OperatorContext getOperatorContext()
+    {
+        return operatorContext;
+    }
+
+    @VisibleForTesting
+    public State getState()
+    {
+        return state;
+    }
+
+    @Override
+    public ListenableFuture<?> isBlocked()
+    {
+        switch (state) {
+            case CLOSED:
+            case CONSUMING_INPUT:
+                return NOT_BLOCKED;
+
+            case LOOKUP_SOURCE_BUILT:
+                return lookupSourceNotNeeded.orElseThrow(() -> new IllegalStateException("Lookup source built, but disposal future not set"));
+        }
+        throw new IllegalStateException("Unhandled state: " + state);
+    }
+
+    @Override
+    public boolean needsInput()
+    {
+        boolean stateNeedsInput = (state == State.CONSUMING_INPUT);
+        return stateNeedsInput && !lookupSourceFactoryDestroyed.isDone();
+    }
+
+    @Override
+    public void addInput(Page page)
+    {
+        requireNonNull(page, "page is null");
+
+        if (lookupSourceFactoryDestroyed.isDone()) {
+            close();
+            return;
+        }
+
+        checkState(state == State.CONSUMING_INPUT);
+
+        layoutSpecificAriaHash.addInput(page);
+    }
+
+    private void updateIndex(Page page)
+    {
+        index.addPage(page);
+
+        if (!localUserMemoryContext.trySetBytes(index.getEstimatedSize().toBytes())) {
+            index.compact();
+            localUserMemoryContext.setBytes(index.getEstimatedSize().toBytes());
+        }
+
+        operatorContext.recordOutput(page.getSizeInBytes(), page.getPositionCount());
+    }
+
+    @Override
+    public void finishMemoryRevoke()
+    {
+        checkState(finishMemoryRevoke.isPresent(), "Cannot finish unknown revoking");
+        finishMemoryRevoke.get().run();
+        finishMemoryRevoke = Optional.empty();
+    }
+
+    @Override
+    public Page getOutput()
+    {
+        return null;
+    }
+
+    @Override
+    public void finish()
+    {
+        if (lookupSourceFactoryDestroyed.isDone()) {
+            close();
+            return;
+        }
+
+        if (finishMemoryRevoke.isPresent()) {
+            return;
+        }
+
+        switch (state) {
+            case CONSUMING_INPUT:
+                finishInput();
+                return;
+
+            case LOOKUP_SOURCE_BUILT:
+                disposeLookupSourceIfRequested();
+                return;
+
+            case CLOSED:
+                // no-op
+                return;
+        }
+
+        throw new IllegalStateException("Unhandled state: " + state);
+    }
+
+    private void finishInput()
+    {
+        checkState(state == State.CONSUMING_INPUT);
+        if (lookupSourceFactoryDestroyed.isDone()) {
+            close();
+            return;
+        }
+
+        LookupSourceSupplier partition = buildLookupSource();
+        localUserMemoryContext.setBytes(partition.get().getInMemorySizeInBytes());
+        lookupSourceNotNeeded = Optional.of(lookupSourceFactory.lendPartitionLookupSource(partitionIndex, partition));
+
+        state = State.LOOKUP_SOURCE_BUILT;
+    }
+
+    private void disposeLookupSourceIfRequested()
+    {
+        checkState(state == State.LOOKUP_SOURCE_BUILT);
+        verify(lookupSourceNotNeeded.isPresent());
+        if (!lookupSourceNotNeeded.get().isDone()) {
+            return;
+        }
+
+        index.clear();
+        localRevocableMemoryContext.setBytes(0);
+        localUserMemoryContext.setBytes(index.getEstimatedSize().toBytes());
+        close();
+    }
+
+    private LookupSourceSupplier buildLookupSource()
+    {
+        return layoutSpecificAriaHash.createLookupSourceSupplier(operatorContext.getSession(), hashChannels, preComputedHashChannel, filterFunctionFactory, sortChannel, searchFunctionFactories, Optional.of(outputChannels));
+    }
+
+    @Override
+    public boolean isFinished()
+    {
+        if (lookupSourceFactoryDestroyed.isDone()) {
+            // Finish early when the probe side is empty
+            close();
+            return true;
+        }
+
+        return state == State.CLOSED;
+    }
+
+    @Override
+    public void close()
+    {
+        if (state == State.CLOSED) {
+            return;
+        }
+        // close() can be called in any state, due for example to query failure, and must clean resource up unconditionally
+
+        state = State.CLOSED;
+        finishMemoryRevoke = finishMemoryRevoke.map(ifPresent -> () -> {});
+
+        try (Closer closer = Closer.create()) {
+            closer.register(index::clear);
+            closer.register(() -> localUserMemoryContext.setBytes(0));
+            closer.register(() -> localRevocableMemoryContext.setBytes(0));
+        }
+        catch (IOException e) {
+            throw new RuntimeException(e);
+        }
+    }
+
+    @Override
+    public boolean retainsInputPages()
+    {
+        // Aria copies build side pages, Presto retains them in index.
+        return layoutSpecificAriaHash == null;
+    }
+}
